Texts of Import Score.ini Files
Chinese and Russian texts of Import Score.ini Files. Here I want to tell something about Russian countings:
Russian countings do not work the same like English, French and Spanish - the countings are only determined by the last digits:
Ends in 1 do not change the case in sentence;
Ends from 2 to 4 use Genitive (or 2nd) case singular(-а/я, -ы/и);
Ends in 0, 5 to 9, or 11 to 19 use Genitive plural(-ей/ев/ёв/ов, or any consonant or a soft sign).
There is only one counting here that defaultly uses the Genitive case, so we may safely use this case and its plural without using the dumb method of word stem + dot.

By the way, the Russian text revisions, include:
Russian uses a dash symbol (–) to show conjugation of verb to be in English (or verb être in French).
Messes of using the word персонаж (character).
Other revisions are inconvenient to describe.

I will continue to do my best on Chinese and Russian text, as previously said, and I will support the completion of Japanese, French, Spanish, Dutch and Korean text.