Unlockable Conditions in Singular Form
It needs to provide unlockable conditions in singular form for English. This could be also helpful to French, Spanish, Dutch, German, Russian and other European languages. Therefore Japanese, Chinese and Korean only moved a few as East Asian Languages.
It is very important to provide singular form, even if the users can comprehend before this. My software testing textbook couldn’t ignore this in L12N module. Some might not tend to be used, like UNLOCK_CONDITION_EARN_SINGULAR, but it does not mean there is no one may use this.
In Russian, singular omits the quantity in default as it does not have articles like a/an and the (or le/la/l’ and les in French). There is no intention to solve casing problems in this time.