Create linux_installer_standalone.bash
This is supposed to be tagged as release and go directly into release (for installing the latest version)

Over the last weeks, I've been trying to add an easy to use OpenTaiko installer to the Arch User Repository (AUR)
This was without success, because OpenTaiko uses many shift_jis encoded files, which special characters, that cannot be taken by the AUR packaging system.

This installer is mobile, and should run from everywhere on Linux. 
As for the compatibility, we should keep a well-known/working list.